Move parserstate.c & .h files into parser.c and .h
