+++ include/rbs/parser.h
@@ -1,7 +1,170 @@
 #ifndef RBS__PARSER_H
 #define RBS__PARSER_H
 
-#include "parserstate.h"
+#include "rbs/defines.h"
+#include "rbs/util/rbs_allocator.h"
+#include "rbs/util/rbs_constant_pool.h"
+#include "rbs/lexer.h"
+#include "rbs/ast.h"
+
+#include <stdbool.h>
+#include <stddef.h>
+
+/**
+ * id_table represents a set of RBS constant IDs.
+ * This is used to manage the set of bound variables.
+ * */
+typedef struct id_table {
+  size_t size;
+  size_t count;
+  rbs_constant_id_t *ids;
+  struct id_table *next;
+} id_table;
+
+/**
+ * comment represents a sequence of comment lines.
+ *
+ *     # Comment for the method.
+ *     #
+ *     # ```rb
+ *     # object.foo()  # Do something
+ *     # ```
+ *     #
+ *     def foo: () -> void
+ *
+ * A comment object represents the six lines of comments.
+ * */
+typedef struct comment {
+  position start;
+  position end;
+
+  size_t line_size;
+  size_t line_count;
+  token *tokens;
+
+  struct comment *next_comment;
+} comment;
+
+typedef struct error {
+  char *message;
+  token token;
+  bool syntax_error;
+} error;
+
+/**
+ * An RBS parser is a LL(3) parser.
+ * */
+typedef struct {
+  lexstate *lexstate;
+
+  token current_token;
+  token next_token;       /* The first lookahead token */
+  token next_token2;      /* The second lookahead token */
+  token next_token3;      /* The third lookahead token */
+
+  id_table *vars;         /* Known type variables */
+  comment *last_comment;  /* Last read comment */
+
+  rbs_constant_pool_t constant_pool;
+  rbs_allocator_t allocator;
+  error *error;
+} parserstate;
+
+/**
+ * Insert new table entry.
+ * Setting `reset` inserts a _reset_ entry, which stops searching.
+ *
+ * ```
+ * class Foo[A]
+ *          ^^^                      <= push new table with reset
+ *   def foo: [B] () -> [A, B]
+ *            ^^^                    <= push new table without reset
+ *
+ *   class Baz[C]
+ *            ^^^                    <= push new table with reset
+ *   end
+ * end
+ * ```
+ * */
+id_table *parser_push_typevar_table(parserstate *state, bool reset);
+NODISCARD bool parser_pop_typevar_table(parserstate *state);
+
+/**
+ * Insert new type variable into the latest table.
+ * */
+NODISCARD bool parser_insert_typevar(parserstate *state, rbs_constant_id_t id);
+
+/**
+ * Returns true if given type variable is recorded in the table.
+ * If not found, it goes one table up, if it's not a reset table.
+ * Or returns false, if it's a reset table.
+ * */
+bool parser_typevar_member(parserstate *state, rbs_constant_id_t id);
+
+/**
+ * Allocate new lexstate object.
+ *
+ * ```
+ * VALUE string = rb_funcall(buffer, rb_intern("content"), 0);
+ * alloc_lexer(string, 0, 31)    // New lexstate with buffer content
+ * ```
+ * */
+lexstate *alloc_lexer(rbs_allocator_t *, rbs_string_t string, const rbs_encoding_t *encoding, int start_pos, int end_pos);
+
+/**
+ * Allocate new parserstate object.
+ * Optionally call `rbs_parser_declare_type_variables_from_ruby_array()` after, to populate the type variable table.
+ *
+ * Once allocated, optionally call `rbs_parser_declare_type_variables` if you'd like to declare
+ * any type variables to be used during the parsing.
+ *
+ * ```
+ * alloc_parser(buffer, string, encoding, 0, 1);
+ * rbs_parser_declare_type_variables_from_ruby_array(variables);
+ * ```
+ * */
+parserstate *alloc_parser(rbs_string_t string, const rbs_encoding_t *encoding, int start_pos, int end_pos);
+void free_parser(parserstate *parser);
+
+/**
+ * Declare type variables to be used during parsing.
+ * */
+void rbs_parser_declare_type_variables(parserstate *parser, size_t count, const char **variables);
+
+/**
+ * Advance one token.
+ * */
+void parser_advance(parserstate *state);
+
+/**
+ * Advance one token if the next_token is a token of the type.
+ * */
+bool parser_advance_if(parserstate *state, enum RBSTokenType type);
+void print_parser(parserstate *state);
+
+/**
+ * Insert new comment line token.
+ * */
+void insert_comment_line(parserstate *state, token token);
+
+/**
+ * Returns a RBS::Comment object associated with an subject at `subject_line`.
+ *
+ * ```rbs
+ * # Comment1
+ * class Foo           # This is the subject line for Comment1
+ *
+ *   # Comment2
+ *   %a{annotation}    # This is the subject line for Comment2
+ *   def foo: () -> void
+ * end
+ * ```
+ * */
+rbs_ast_comment_t *get_comment(parserstate *state, int subject_line);
+
+comment *alloc_comment(rbs_allocator_t *, token comment_token, comment *last_comment);
+void comment_insert_new_line(rbs_allocator_t *, comment *com, token comment_token);
+comment *comment_get_comment(comment *com, int line);
 
 void set_error(parserstate *state, token tok, bool syntax_error, const char *fmt, ...) RBS_ATTRIBUTE_FORMAT(4, 5);
